requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
packaging==24.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
virtualenv==20.26.2
Werkzeug==3.0.2